refactor: rename `pdf` to `structure`
